test: consolidate native Qwen3.5 model coverage
