--- tests/unit/train/models/test_qwen3_5_moe.py
@@ -1,175 +1,0 @@
-import pytest
-import torch
-
-from prime_rl.configs.trainer import ModelConfig
-from prime_rl.trainer.model import resolve_auto_attn
-from prime_rl.trainer.models import AutoModelForCausalLMPrimeRL
-from prime_rl.trainer.models.fusions import apply_model_fusions
-from prime_rl.trainer.models.layers.lm_head import inject_prime_lm_head
-from prime_rl.trainer.models.qwen3_5 import Qwen3_5ForCausalLM, Qwen3_5MoeTextConfig
-from prime_rl.utils.cp import setup_model_cp
-from prime_rl.utils.utils import default_dtype
-
-pytestmark = [pytest.mark.gpu]
-
-
-def get_model():
-    runtime_config = ModelConfig()
-    resolve_auto_attn(runtime_config)
-    config = Qwen3_5MoeTextConfig(
-        vocab_size=256,
-        hidden_size=256,
-        num_hidden_layers=4,
-        num_attention_heads=4,
-        num_key_value_heads=2,
-        head_dim=64,
-        moe_intermediate_size=128,
-        shared_expert_intermediate_size=128,
-        num_experts=8,
-        num_experts_per_tok=2,
-        max_position_embeddings=512,
-        rms_norm_eps=1e-6,
-        linear_conv_kernel_dim=4,
-        linear_key_head_dim=32,
-        linear_value_head_dim=32,
-        linear_num_key_heads=4,
-        linear_num_value_heads=8,
-    )
-    with torch.device("cuda"):
-        prime_model = AutoModelForCausalLMPrimeRL.from_config(
-            config, attn_implementation=runtime_config.attn, dtype=torch.bfloat16
-        )
-    inject_prime_lm_head(prime_model, chunk_size=None)
-    return prime_model
-
-
-def test_qwen3_5_moe():
-    prime_model = get_model()
-    apply_model_fusions(prime_model, ["qkv", "gate_up"])
-    input_ids = torch.randint(0, prime_model.config.vocab_size, (1, 100), device="cuda")
-    position_ids = torch.arange(1, 101, device="cuda").unsqueeze(0)
-    prime_output = prime_model(
-        input_ids,
-        position_ids=position_ids,
-        seq_lens=torch.tensor([input_ids.shape[1]], device="cuda"),
-    )
-    prime_output["logits"].sum().backward()
-    assert torch.isfinite(prime_output["logits"]).all()
-    assert torch.isfinite(prime_model.model.embed_tokens.weight.grad).all()
-
-    packed_position_ids = torch.arange(1, 51, device="cuda").repeat(2).unsqueeze(0)
-    # Keep expert selection fixed to isolate packed sequence boundaries.
-    config = prime_model.config
-    routed_experts = (
-        torch.rand(1, 100, config.num_hidden_layers, config.num_experts, device="cuda")
-        .topk(config.num_experts_per_tok, dim=-1)
-        .indices
-    )
-    with torch.no_grad():
-        packed = prime_model(
-            input_ids,
-            position_ids=packed_position_ids,
-            seq_lens=torch.tensor([50, 50], device="cuda"),
-            routed_experts=routed_experts,
-        )["logits"]
-        unpacked = torch.cat(
-            [
-                prime_model(
-                    input_ids[:, start : start + 50],
-                    position_ids=packed_position_ids[:, :50],
-                    seq_lens=torch.tensor([50], device="cuda"),
-                    routed_experts=routed_experts[:, start : start + 50],
-                )["logits"]
-                for start in (0, 50)
-            ],
-            dim=1,
-        )
-    torch.testing.assert_close(packed, unpacked, atol=0.03, rtol=0.01)
-
-
-def test_qwen3_5_moe_router_replay():
-    """When routed_experts are provided, the model uses them instead of computing routing."""
-    prime_model = get_model()
-
-    with torch.device("cuda"), default_dtype(torch.bfloat16):
-        input_ids = torch.randint(0, prime_model.config.vocab_size, (1, 100))
-        position_ids = torch.arange(1, 101).unsqueeze(0)
-
-    seq_lens = torch.tensor([input_ids.shape[1]], device="cuda")
-    out_normal = prime_model(input_ids, position_ids=position_ids, seq_lens=seq_lens)
-
-    num_layers = prime_model.config.num_hidden_layers
-    topk = prime_model.config.num_experts_per_tok
-    routed_experts = torch.randint(0, prime_model.config.num_experts, (1, 100, num_layers, topk), device="cuda")
-
-    prime_model.zero_grad()
-    out_replay = prime_model(
-        input_ids,
-        position_ids=position_ids,
-        routed_experts=routed_experts,
-        seq_lens=seq_lens,
-    )
-
-    assert out_replay["logits"].shape == out_normal["logits"].shape
-
-    out_replay["logits"].sum().backward()
-    assert prime_model.model.embed_tokens.weight.grad is not None
-
-
-def test_qwen3_5_moe_cp_patching():
-    from unittest.mock import MagicMock
-
-    from prime_rl.trainer.models.afmoe.modeling_afmoe import AfmoeFlashAttention
-    from prime_rl.trainer.models.layers.attn import FlashAttention, substitute_ring_attn
-    from prime_rl.trainer.models.qwen3_5.attention import Qwen3_5Attention
-
-    originals = {cls: cls._compute_attention for cls in (FlashAttention, AfmoeFlashAttention)}
-    try:
-        mock_group = MagicMock()
-        substitute_ring_attn(process_group=mock_group, heads_k_stride=1)
-        assert Qwen3_5Attention._compute_attention is FlashAttention._compute_attention
-        assert Qwen3_5Attention._compute_attention is not originals[FlashAttention]
-    finally:
-        for cls, method in originals.items():
-            cls._compute_attention = method
-
-
-def test_qwen3_5_moe_context_parallel_setup_hook():
-    from unittest.mock import MagicMock
-
-    config = Qwen3_5MoeTextConfig(
-        vocab_size=128,
-        hidden_size=64,
-        num_hidden_layers=2,
-        layer_types=["linear_attention", "full_attention"],
-        num_attention_heads=4,
-        num_key_value_heads=2,
-        head_dim=16,
-        moe_intermediate_size=64,
-        shared_expert_intermediate_size=64,
-        num_experts=4,
-        num_experts_per_tok=2,
-        max_position_embeddings=128,
-        linear_conv_kernel_dim=4,
-        linear_key_head_dim=8,
-        linear_value_head_dim=8,
-        linear_num_key_heads=4,
-        linear_num_value_heads=8,
-    )
-    with torch.device("meta"):
-        model = Qwen3_5ForCausalLM(config)
-
-    linear_layer = model.model.layers[0]
-    model.model.layers[0] = torch.nn.Sequential(linear_layer)
-    cp_group = MagicMock()
-    setup_model_cp(model, cp_group, cp_rank=1, cp_world_size=2)
-
-    assert model.model.context_parallel_group is cp_group
-    assert model.model.context_parallel_rank == 1
-    assert model.model.context_parallel_world_size == 2
-    assert linear_layer.linear_attn.context_parallel_group is cp_group
-    assert linear_layer.linear_attn.context_parallel_world_size == 2
-
-
-if __name__ == "__main__":
-    test_qwen3_5_moe()
